requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
django-cleanup==8.0.0
django-filter==23.5
django-widget-tweaks==1.5.0
